Removed ModTime checks
This feature was only added to allow redo to skip running md5sum on big files.
The data integrety guarantees of redo are too important to give up for a small optimization.
Also, I suspect this will not play with using directories as $3 in do executables.
Hashing is simpler and safer.