Fixed some logic for ifcreate and unless-change
The logic for these was just wrong in some cases.
It's tricky making sure that all dependencies are handled correctly
whether this is the first time building the dep tree or the nth time.
This involved improving code in NewNode to better distinguish
between all the possible kinds of dependencies that we can talk about.